requirements.txt
matplotlib==3.4.1
natsort==7.1.1
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.2.4
Pillow==8.2.0
pytorch-msssim==0.2.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.7.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tifffile==2021.4.8
torch==1.8.1
torch-optimizer==0.1.0
torchfile==0.1.0
torchmetrics==0.3.1
torchvision==0.9.1
tqdm==4.60.0
visdom==0.1.8.9